LayoutTests/fast/js/regexp-test-null-string-expected.txt

This page tests the exception you get when you test a regular expression against the null string. If the test passes, you'll see a PASS message below.

PASS: e should be 'Error: No input to /a/.' and is.
